duality_offroad_autonomy/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch==0.3.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
albumentations>=1.3.1
seaborn>=0.13.0
pandas>=2.0.0
